sdk/python/requirements.txt
requests==2.32.4
nats-py==2.10.0
httpx==0.28.1
